+++ Changelog.md
@@ -6,6 +6,7 @@
 * Ignore invalid `diaspora://` links [#7652](https://github.com/diaspora/diaspora/pull/7652)
 
 ## Features
+* Add birthday notifications [#7624](https://github.com/diaspora/diaspora/pull/7624)
 
 # 0.7.1.1